Jumbo commit for new applet loading policies.
Applets are now loaded in a "group" which is normally keyed off the
server containing the page that contains the applet (this is often,
but not necessarily, the server containing the applet!).

There will be a preferences dialog to change the group assignments,
but I hesitate to check it in since Ken has threatened to rewrite
the framework.