FormInfo.InputRadio.setup(): By commenting out the call of self.var.set(), unchecked radio arrays act like they do in Netscape and Mosaic namely:
1. initially no radios in the array are selected

2. once a radio has been selected, there's no way to just unselected
   it, short of resetting the entire form

3. upon `reload' of the page, any selected radio stays selected.
   reload does *not* reset the radio.

4. using the form's reset button, all radios are unselected.

FormInfo.InputRadio.reset(): make sure that upon reset, if no radios
are checked, none will be selected.

Note that this is counter to the HTML 2.0 spec, section 8.1.2.4, but
because of Netscape and Mosaic behavior, it's what people expect.  Is
the spec broken?  You decide!

There is no preference or option for pedantic adherence to the HTML
2.0 spec.