Improve performance of CacheItem.getdata by maintaining a list of strings rather than a single string to represent the date returned by the API.
The assumed common case is that multiple readers will be asking for
the same sized chunks of data with getdata. Therefore, it stores a
dictionary of offset -> index into list of strings. This makes for
fast lookup in the default case, where the extra readers are getting
the same sized chunks of data.

In the worst case, a second reader will ask for smaller chunks of data
with getdata. As a result, all its calls to getdata will use different
offsets than the first call and the dictionary will be useless; I fall
back to a linear search of the keys, realizing that this provides (!)
O(k) lookup, where k is the number of keys, on a single call to
getdata and O(k^2) reading of an entire page, where k is the number of
chunks.