Fixed a particularly nasty bug reproducable by:
1. Start up Grail as normal
2. Create a new browser with Alt-N
3. Close the *first* browser window with Alt-W
4. Bring up the Bookmarks Dialog with Alt-B

Grail will barf because the Tk window of the first (now deleted)
browser window was being used as the root of Bookmarks Dialog.

This fix makes Application.root be the master of all the dialogs
created in the Bookmarks module.  Also
BookmarksController.get_browser()/set_browser() is smarter about
finding a browser window to launch on.