Corrected font sizes mistakes (all who are design/layout sensitive please evaluate, so we can come to a good baseline!):
 - "medium" had the same size for h1 and h2!?  I bumped up the h1 a bit.
 - "huge" had h2 and h3 swapped - whoops.

Guido has mentioned that some of the different sizes actually show the
same size fonts for particular families on the SGI - perhaps it was
the times/large 20 and 18, i don't recall.  Anyway, suggestions as to
settings that work better for your X font repertoire/sensibilities are
welcome, i'll try to come up with the best compromise i can.