Licencing things:
- At the top of the file there's a 3-line copyright notice, pointing
to the license by handle, by URL and by filename.  This will be
repeated in each file:

# Copyright (c) CNRI 1996, licensed under terms and conditions of license
# agreement obtained from handle "hdl:CNRI.License/Grail-Version-0.3",
# URL "http://grail.cnri.reston.va.us/LICENSE-0.3/", or file "LICENSE".

- Changed the notice displayed in the splash screen to reflect the
copyright and trademark status:

	Copyright \251 1996.
	GRAIL\256, is a registered
	trademark of CNRI.

(\251 is a C in a circle, \256 is an R in a circle.)

The copyright lists 1996 because no "copy" (i.e. hardcopy, like
CD-ROM) has been published in 1997 yet.  We have the CD-ROM in
Programming Python by Mark Lutz and in the Red Hat 4.0 distribution as
copies distributed in 1996.


The business with the splash screen and the displaying of the license
has to be changed again.  Here's what I agreed with Al Vezza:
When Grail comes up initially, it displays a "default internal home
page" which displays the copyright and trademark status and pointers
to the license by handle, URL and filename.  The user can change this
default home page with the preferences panel.  I suggest to make it a
nice page with some warm welcoming words, links to the real Grail and
Python home page, etc.  As Ken suggested, the link to the license will
also be added to the Help menu.  The initial, crummy license dialog
will be removed.  I think I'll also get rid of the splash screen --
the internal home page can contain that image, and the "About Grail"
help menu item can point to it (rather than to an on-line web page).

I think I'll have to introduce a new URL scheme (similar to "doc:",
which is about 10 lines of code) to have a clean way to refer to the
internal home page.  No big deal.