More work towards the spec.  I read the hash table spec and got the standard / example hash table from the location indicated in the spec:
ftp://cnri.reston.va.us/handles/client_library/hdl_hash.tbl;type=i

and reworked the HashTable logic so it actually worked with this
version (it is schema version 1, not 2, but the bucket format appears
unchanged).  This resulted in several more handles being resolved!
(Though some handles now refer to a handle server that appears down.)

The new hash table format starts with a 16-byte MD5 checksum, instead
of an 11-byte (sic) magic word.  If the checksum is bad, the file is
ignored as if it did not exist.

All four buckets from the standard hash table are now used for the
hardcoded fallback scheme.

I changed and added several debug statements.  The hash bucket
contents is now printed when it is first read, not each time it is
encountered.


I changed the default and fallback hash table filenames to end in
".tbl" instead of "_tbl", to match the standard / example file.

I removed a longish comment containing a class HandleCache, which
seemed to consist of a pie-in-the-sky docstring only.

I also nixed the unused HSConfig class, which seemed to be getting the
same info that's in the hash table anyway (and since I could find no
refs to it in the spec not any live examples, I expect that its use is
discontinued).