Here's a fix for a nasty condition.  I'm not entirely sure that the fix doesn't break anything else, so please beware.
Here's the condition: when an error occurs resolving a handle, the
pollmeta() raises IOError.  This happens after BaseReader.__init__ has
already registered itself as a reader, but as the exception happens in
the constructor, this prevents the corresponding rmreader() call to be
made.

This problem only occurs when IOError is raised by pollmeta() -- for
most other protocols, it is raised by the protocol's __init__(), and
in that case the reader registration hasn't happened yet.

Here's the fix: I surround the self.start() call (which is the
culprit, albeit rather indirectly) with a try-finally which calls
rmreader() when the self.start() call did not return.  A local flag
variable is used to keep track of this.