Use the new hdllib functions fetch_global_hash_table() and fetch_local_hash_table().  We are now no longer dependent on a correct hash table file, and we can resolve handles via local handle servers if needed.  The global hash table and the local hash tables (indexed by authority name) are saved as global variables in the module.
We don't yet try to find a caching local handle server first, and we
don't reload hash tables when we get the error indicating that we
should, etc., but we should be good for a while now!



This has implications for Roj's specially modified hdlAPI.py.  The
special hdlAPI.py is no longer needed to resolve handles in the
nlm.hdl_test domain to URLs -- but it may still be necessary to
resolve them to HDL_TYPE_DLS data.

Perhaps we can design an extension to hdlAPI that interprets the
";type=..." syntax and attempts to dynamically load a support module
for that data type; this would avoid having two copies of hdlAPI.py
around.