Fred: I decided to code this up after all :-)
Changed the way to customize the help menu, doing away with the extra
indirections.  There's now a preference browser--help-menu whose
contents is split at \n characters, and lines are alternatingly labels
and URLs for the help menu.  A hyphen means a separator as before (and
has no URL).

Also changed the name of landmarks--about-grail-page to
landmarks--default-home-page (which is closer to its meaning).

Wait for the accompanying checkin of grail-defaults.