Enhance the 'custom defined' extensibility for 'hdl' protocol handling.  In this case
By defining a function called 'exception_handler', in a module called hdl_errors,
in the users personal 'protocols' directory; exception_handler() is called
instead of an IOError being raised.  In this way, the developer is given control over
the error handling behavior associated with a failed handle lookup.  For example
if the handle is not found, she may not wish the grail error dialog to appear, but
rather try and resolve the error through other means.

exception_handler must raise IOError or correctly return the (replyflags, _items)
tuple.  Absents of this module or function (or if it can be imported)
causes grail to silently handle errors in the default way by raising IOError.